ref: move listItemUpdate controller to controllers.publish.ts
