+++ src/server/components/dashboard/listsItems/controllers.ts
@@ -143,29 +143,6 @@ export async function listItemPostController(req: Request, res: Response) {
   return res.redirect(`${listItemUrl}/confirm`);
 }
 
-export async function listItemUpdateController(req: Request, res: Response): Promise<void> {
-  const userId = req.user!.id;
-  const listItem = res.locals.listItem;
-  const { listItemUrl, listIndexUrl } = res.locals;
-
-  try {
-    await handleListItemUpdate(listItem.id, userId);
-
-    if (userId === undefined) {
-      req.flash("errorMsg", "Unable to perform action - user could not be identified");
-      return res.redirect(listItemUrl);
-    }
-
-    req.flash("successBannerTitle", `${listItem.jsonData.organisationName} has been updated and published`);
-    req.flash("successBannerHeading", "Updated and published");
-    req.flash("successBannerColour", "green");
-    return res.redirect(listIndexUrl);
-  } catch (error: any) {
-    req.flash("errorMsg", `${listItem.jsonData.organisationName} could not be updated. ${error.message}`);
-    return res.redirect(listItemUrl);
-  }
-}
-
 export async function handleListItemUpdate(id: number, userId: User["id"]): Promise<void> {
   logger.info(`${userId} looking for ${id} to update`);
   const listItem = await prisma.listItem.findUnique({
